+++ src/Services/Main/Result/EventLogItemResult.php
@@ -13,7 +13,9 @@ declare(strict_types=1);
 
 namespace Bitrix24\SDK\Services\Main\Result;
 
+use Bitrix24\SDK\Attributes\OaEntity;
 use Bitrix24\SDK\Core\Result\AbstractItem;
+use Bitrix24\SDK\Services\Main\Service\EventLogSelectBuilder;
 use Carbon\CarbonImmutable;
 use Darsyn\IP\Version\Multi;
 
@@ -32,6 +34,10 @@ use Darsyn\IP\Version\Multi;
  * @property-read int|null             $guestId
  * @property-read string|null          $description
  */
+#[OaEntity(
+    entityKey:     'bitrix.main.eventlogdto',
+    selectBuilder: EventLogSelectBuilder::class,
+)]
 class EventLogItemResult extends AbstractItem
 {
     /**